fix breaking change - abstract get neighbours function
